Move metadata async tests out of production module
